requirements.txt
bcrypt
PyJWT
pyotp